refactor(dpp): version compute_minimum_shielded_fee with the mod + v0 pattern
Move the shielded minimum-fee formula behind the project's standard platform-versioned
dispatch (mirroring daily_withdrawal_limit / registration_cost):

- packages/rs-dpp/src/shielded/compute_minimum_shielded_fee/mod.rs dispatches on the new
  `platform_version.dpp.methods.compute_minimum_shielded_fee` FeatureVersion, with an
  `UnknownVersionMismatch` fallback.
- The v0 formula (the checked-arithmetic implementation) lives in
  compute_minimum_shielded_fee/v0/mod.rs.
- Added the `compute_minimum_shielded_fee` FeatureVersion field to DPPMethodVersions
  (initialized to 0 in both DPP_METHOD_VERSIONS_V1 and _V2).

The public path `dpp::shielded::compute_minimum_shielded_fee` is preserved via re-export
(the module and the function share a name but live in different namespaces), so no call
site changes were needed (builders, transformers, the consensus validator, the wallet).

Verified: clippy --all-features --all-targets -D warnings clean; dpp shielded::builder 25,
drive-abci validate_minimum_shielded_fee 6 pass, 0 failures.
