+++ packages/rs-platform-version/src/version/dpp_versions/dpp_method_versions/mod.rs
@@ -8,4 +8,5 @@ pub struct DPPMethodVersions {
     pub epoch_core_reward_credits_for_distribution: FeatureVersion,
     pub daily_withdrawal_limit: FeatureVersion,
     pub deduct_fee_from_outputs_or_remaining_balance_of_inputs: FeatureVersion,
+    pub compute_minimum_shielded_fee: FeatureVersion,
 }
