+++ packages/rs-platform-version/src/version/dpp_versions/dpp_method_versions/v1.rs
@@ -3,4 +3,5 @@ pub const DPP_METHOD_VERSIONS_V1: DPPMethodVersions = DPPMethodVersions {
     epoch_core_reward_credits_for_distribution: 0,
     daily_withdrawal_limit: 0,
     deduct_fee_from_outputs_or_remaining_balance_of_inputs: 0,
+    compute_minimum_shielded_fee: 0,
 };
